[FIRRTL][HW] Move HierPathOp from FIRRTL to HW
Move the HierPathOp from the FIRRTL Dialect to the HW Dialect.  This is
done in preparation to replace the GlobalRefOp with HierPathOp.
