+++ include/circt/Dialect/FIRRTL/FIRRTLStructure.td
@@ -200,79 +200,3 @@ def FMemModuleOp : FIRRTLModuleLike<"memmodule"> {
 
   let hasCustomAssemblyFormat = 1;
 }
-
-def HierPathOp : FIRRTLOp<"hierpath",
-      [IsolatedFromAbove, Symbol,
-       DeclareOpInterfaceMethods<InnerRefUserOpInterface>,
-       HasParent<"CircuitOp">]> {
-  let summary = "Hierarchical path specification";
-  let description = [{
-    The "firrtl.hierpath" operation represents a path through the hierarchy.
-    This is used to specify namable things for use in other operations, for
-    example in verbatim substitution. Non-local annotations also use these.
-  }];
-  let arguments = (ins SymbolNameAttr:$sym_name, NameRefArrayAttr:$namepath);
-  let results = (outs);
-  let hasCustomAssemblyFormat = 1;
-  let extraClassDeclaration = [{
-    /// Drop the module from the namepath. If its a InnerNameRef, then drop
-    /// the Module-Instance pair, else drop the final module from the namepath.
-    /// Return true if any update is made.
-    bool dropModule(StringAttr moduleToDrop);
-
-    /// Inline the module in the namepath.
-    /// Update the symbol name for the inlined module instance, by prepending
-    /// the symbol name of the instance at which the inling was done.
-    /// Return true if any update is made.
-    bool inlineModule(StringAttr moduleToDrop);
-
-    /// Replace the oldMod module with newMod module in the namepath of the NLA.
-    /// Return true if any update is made.
-    bool updateModule(StringAttr oldMod, StringAttr newMod);
-
-    /// Replace the oldMod module with newMod module in the namepath of the NLA.
-    /// Since the module is being updated, the symbols inside the module should
-    /// also be renamed. Use the rename Map to update the corresponding
-    /// inner_sym names in the namepath. Return true if any update is made.
-    bool updateModuleAndInnerRef(StringAttr oldMod, StringAttr newMod,
-         const llvm::DenseMap<StringAttr, StringAttr> &innerSymRenameMap);
-
-    /// Truncate the namepath for this NLA, at atMod module.
-    /// If includeMod is false, drop atMod and beyond, else include it and drop
-    /// everything after it.
-    /// Return true if any update is made.
-    bool truncateAtModule(StringAttr atMod, bool includeMod = true);
-
-    /// Return just the module part of the namepath at a specific index.
-    StringAttr modPart(unsigned i);
-
-    /// Return the root module.
-    StringAttr root();
-
-    /// Return just the reference part of the namepath at a specific index.
-    /// This will return an empty attribute if this is the leaf and the leaf is
-    /// a module.
-    StringAttr refPart(unsigned i);
-
-    /// Return the leaf reference.  This returns an empty attribute if the leaf
-    /// reference is a module.
-    StringAttr ref();
-
-    /// Return the leaf Module.
-    StringAttr leafMod();
-
-    /// Returns true, if the NLA path contains the module.
-    bool hasModule(StringAttr modName);
-
-    /// Returns true, if the NLA path contains the InnerSym {modName, symName}.
-    bool hasInnerSym(StringAttr modName, StringAttr symName) const;
-
-    /// Returns true if this NLA targets a module or instance of a module (as
-    /// opposed to an instance's port or something inside an instance).
-    bool isModule();
-
-    /// Returns true if this NLA targets something inside a module (as opposed
-    /// to a module or an instance of a module);
-    bool isComponent();
-    }];
-}
