+++ test/Dialect/FIRRTL/emit-omir.mlir
@@ -206,7 +206,7 @@ firrtl.circuit "NonLocalTrackers" attributes {annotations = [{
 }]} {
   // Both OMReferenceTarget1 and OMReferenceTarget2 share the same NLA.  This
   // NLA should not be deleted.
-  firrtl.hierpath private @nla_0 [@NonLocalTrackers::@b, @B::@a, @A]
+  hw.hierpath private @nla_0 [@NonLocalTrackers::@b, @B::@a, @A]
   // CHECK: firrtl.module @A
   firrtl.module @A() attributes {annotations = [{circt.nonlocal = @nla_0, class = "freechips.rocketchip.objectmodel.OMIRTracker", id = 0}]} {
     // CHECK-NEXT: %a = firrtl.wire sym @[[a_sym:[^ ]+]]
@@ -326,7 +326,7 @@ firrtl.circuit "Top"  attributes {annotations = [{
     info = #loc4
   }]
 }]} {
-  firrtl.hierpath private @nla [@Top::@a, @A]
+  hw.hierpath private @nla [@Top::@a, @A]
   firrtl.module @Top(in %in0_0: !firrtl.uint<4>, in %in0_1: !firrtl.uint<4>, in %in1_f0: !firrtl.uint<4>, in %in1_f1: !firrtl.uint<4>, out %out0_0: !firrtl.uint<4>, out %out0_1: !firrtl.uint<4>, out %out1_f0: !firrtl.uint<4>, out %out1_f1: !firrtl.uint<4>) {
     %a_in0, %a_in1, %a_out0, %a_out1 = firrtl.instance a sym @a  @A(in in0: !firrtl.bundle<io: vector<uint<4>, 2>>, in in1: !firrtl.bundle<io: bundle<f0: uint<4>, f1: uint<4>>>, out out0: !firrtl.vector<uint<4>, 2>, out out1: !firrtl.bundle<f0: uint<4>, f1: uint<4>>)
     %0 = firrtl.subfield %a_in1[io] : !firrtl.bundle<io: bundle<f0: uint<4>, f1: uint<4>>>
@@ -528,8 +528,8 @@ firrtl.circuit "SRAMPathsWithNLA" attributes {annotations = [{
     }
   ]
 }]} {
-  firrtl.hierpath private @nla_old [@SRAMPathsWithNLA::@s1, @Submodule::@mem]
-  firrtl.hierpath private @nla_new [@SRAMPathsWithNLA::@s1, @Submodule]
+  hw.hierpath private @nla_old [@SRAMPathsWithNLA::@s1, @Submodule::@mem]
+  hw.hierpath private @nla_new [@SRAMPathsWithNLA::@s1, @Submodule]
   firrtl.module @Submodule() {
     %mem_port = firrtl.mem sym @mem Undefined {
       annotations = [
@@ -607,7 +607,7 @@ firrtl.circuit "SRAMPathsWithNLA" attributes {annotations = [{
     }
   ]
 }]} {
-  firrtl.hierpath private @nla [@SRAMPaths::@sub, @Submodule]
+  hw.hierpath private @nla [@SRAMPaths::@sub, @Submodule]
   firrtl.extmodule private @MySRAM()
   firrtl.module private @Submodule() {
     firrtl.instance mem1 {annotations = [{circt.nonlocal = @nla, class = "freechips.rocketchip.objectmodel.OMIRTracker", id = 0}]} @MySRAM()
@@ -935,9 +935,9 @@ firrtl.circuit "FixPath"  attributes {annotations = [
       id = "OMID:0",
       info = loc(unknown)}
     ]}]} {
-  firrtl.hierpath private @nla_3 [@FixPath::@c, @C::@cd, @D]
-  firrtl.hierpath private @nla_2 [@FixPath::@c, @C::@in]
-  firrtl.hierpath private @nla_1 [@FixPath::@c, @C]
+  hw.hierpath private @nla_3 [@FixPath::@c, @C::@cd, @D]
+  hw.hierpath private @nla_2 [@FixPath::@c, @C::@in]
+  hw.hierpath private @nla_1 [@FixPath::@c, @C]
   firrtl.module @C(
     in %in: !firrtl.uint<1> sym @in [
       {circt.nonlocal = @nla_2, class = "freechips.rocketchip.objectmodel.OMIRTracker", id = 1 : i64}
