+++ test/Dialect/FIRRTL/lower-types.mlir
@@ -1113,7 +1113,7 @@ firrtl.module private @is1436_FOO() {
   }
 
   // CHECK-LABEL: firrtl.module @MergeBundle
-  firrtl.module @MergeBundle(out %o: !firrtl.bundle<valid: uint<1>, ready: uint<1>>, in %i: !firrtl.uint<1>) 
+  firrtl.module @MergeBundle(out %o: !firrtl.bundle<valid: uint<1>, ready: uint<1>>, in %i: !firrtl.uint<1>)
   {
     %a = firrtl.wire   : !firrtl.bundle<valid: uint<1>, ready: uint<1>>
     firrtl.strictconnect %o, %a : !firrtl.bundle<valid: uint<1>, ready: uint<1>>
@@ -1126,7 +1126,7 @@ firrtl.module private @is1436_FOO() {
     // CHECK:  firrtl.strictconnect %a_valid, %i : !firrtl.uint<1>
     // CHECK:  firrtl.strictconnect %a_ready, %i : !firrtl.uint<1>
   }
- 
+
   // CHECK-LABEL: firrtl.module @MergeVector
   firrtl.module @MergeVector(out %o: !firrtl.vector<uint<1>, 3>, in %i: !firrtl.uint<1>) {
     %a = firrtl.wire   : !firrtl.vector<uint<1>, 3>
@@ -1161,7 +1161,7 @@ firrtl.circuit "DontTouch" {
 
 // Check that we don't create symbols for non-local annotations.
 firrtl.circuit "Foo"  {
-  firrtl.hierpath private @nla [@Foo::@bar, @Bar]
+  hw.hierpath private @nla [@Foo::@bar, @Bar]
   // CHECK:       firrtl.module private @Bar(in %a_b:
   // CHECK-SAME:    !firrtl.uint<1> [{circt.nonlocal = @nla, class = "circt.test"}])
   firrtl.module private @Bar(in %a: !firrtl.bundle<b: uint<1>>
