+++ test/Dialect/FIRRTL/prefix-modules.mlir
@@ -205,10 +205,10 @@ firrtl.circuit "GCTInterfacePrefix"
 // CHECK: firrtl.circuit "T_NLATop"
 firrtl.circuit "NLATop" {
 
-  firrtl.hierpath private @nla [@NLATop::@test, @Aardvark::@test, @Zebra]
-  firrtl.hierpath private @nla_1 [@NLATop::@test,@Aardvark::@test_1, @Zebra]
-  // CHECK: firrtl.hierpath private @nla [@T_NLATop::@test, @T_Aardvark::@test, @T_A_Z_Zebra]
-  // CHECK: firrtl.hierpath private @nla_1 [@T_NLATop::@test, @T_Aardvark::@test_1, @T_A_Z_Zebra]
+  hw.hierpath private @nla [@NLATop::@test, @Aardvark::@test, @Zebra]
+  hw.hierpath private @nla_1 [@NLATop::@test,@Aardvark::@test_1, @Zebra]
+  // CHECK: hw.hierpath private @nla [@T_NLATop::@test, @T_Aardvark::@test, @T_A_Z_Zebra]
+  // CHECK: hw.hierpath private @nla_1 [@T_NLATop::@test, @T_Aardvark::@test_1, @T_A_Z_Zebra]
   // CHECK: firrtl.module @T_NLATop
   firrtl.module @NLATop()
     attributes {annotations = [{
@@ -306,14 +306,14 @@ firrtl.circuit "GCTDataMemTapsPrefix" {
 // Test the NonLocalAnchor is properly updated.
 // CHECK-LABEL: firrtl.circuit "FixNLA" {
   firrtl.circuit "FixNLA"   {
-    firrtl.hierpath private @nla_1 [@FixNLA::@bar, @Bar::@baz, @Baz]
-    // CHECK:   firrtl.hierpath private @nla_1 [@FixNLA::@bar, @Bar::@baz, @Baz]
-    firrtl.hierpath private @nla_2 [@FixNLA::@foo, @Foo::@bar, @Bar::@baz, @Baz::@s1]
-    // CHECK:   firrtl.hierpath private @nla_2 [@FixNLA::@foo, @X_Foo::@bar, @X_Bar::@baz, @X_Baz::@s1]
-    firrtl.hierpath private @nla_3 [@FixNLA::@bar, @Bar::@baz, @Baz]
-    // CHECK:   firrtl.hierpath private @nla_3 [@FixNLA::@bar, @Bar::@baz, @Baz]
-    firrtl.hierpath private @nla_4 [@Foo::@bar, @Bar::@baz, @Baz]
-    // CHECK:       firrtl.hierpath private @nla_4 [@X_Foo::@bar, @X_Bar::@baz, @X_Baz]
+    hw.hierpath private @nla_1 [@FixNLA::@bar, @Bar::@baz, @Baz]
+    // CHECK:   hw.hierpath private @nla_1 [@FixNLA::@bar, @Bar::@baz, @Baz]
+    hw.hierpath private @nla_2 [@FixNLA::@foo, @Foo::@bar, @Bar::@baz, @Baz::@s1]
+    // CHECK:   hw.hierpath private @nla_2 [@FixNLA::@foo, @X_Foo::@bar, @X_Bar::@baz, @X_Baz::@s1]
+    hw.hierpath private @nla_3 [@FixNLA::@bar, @Bar::@baz, @Baz]
+    // CHECK:   hw.hierpath private @nla_3 [@FixNLA::@bar, @Bar::@baz, @Baz]
+    hw.hierpath private @nla_4 [@Foo::@bar, @Bar::@baz, @Baz]
+    // CHECK:       hw.hierpath private @nla_4 [@X_Foo::@bar, @X_Bar::@baz, @X_Baz]
     // CHECK-LABEL: firrtl.module @FixNLA()
     firrtl.module @FixNLA() {
       firrtl.instance foo sym @foo  @Foo()
@@ -349,10 +349,10 @@ firrtl.circuit "GCTDataMemTapsPrefix" {
 
 // Test that NonLocalAnchors are properly updated with memmodules.
 firrtl.circuit "Test"   {
-  // CHECK: firrtl.hierpath private @nla_1 [@Test::@foo1, @A_Foo1::@bar, @A_Bar]
-  firrtl.hierpath private @nla_1 [@Test::@foo1, @Foo1::@bar, @Bar]
-  // CHECK: firrtl.hierpath private @nla_2 [@Test::@foo2, @B_Foo2::@bar, @B_Bar]
-  firrtl.hierpath private @nla_2 [@Test::@foo2, @Foo2::@bar, @Bar]
+  // CHECK: hw.hierpath private @nla_1 [@Test::@foo1, @A_Foo1::@bar, @A_Bar]
+  hw.hierpath private @nla_1 [@Test::@foo1, @Foo1::@bar, @Bar]
+  // CHECK: hw.hierpath private @nla_2 [@Test::@foo2, @B_Foo2::@bar, @B_Bar]
+  hw.hierpath private @nla_2 [@Test::@foo2, @Foo2::@bar, @Bar]
 
   firrtl.module @Test() {
     firrtl.instance foo1 sym @foo1 @Foo1()
